requirements.txt
asgiref==3.8.1
Django==5.1.1
django-appconf==1.0.6
django-compressor==4.5.1
django-tinymce==4.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.9.0
rcssmin==1.1.2
rjsmin==1.2.2
six==1.16.0
sqlparse==0.5.1
tzdata==2024.2
